requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap3==11.0.0
django-braces==1.13.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.2
enum34==1.1.6
gunicorn==19.6.0
idna==2.8
ipython-genutils==0.1.0
oauthlib==3.0.1
Pillow==5.4.1
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
ptyprocess==0.5.1
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.1.3
requests==2.21.0
requests-oauthlib==1.2.0
simplegeneric==0.8.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
stripe==1.37.0
traitlets==4.3.1
wcwidth==0.1.7
urllib3==1.24.2
whitenoise==3.2.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
